패치노트 10.20
로딩 에셋 범위 추가
-4~5스테이지 몬스터 및 보스 로딩 에셋으로 추가

몬스터 체력 계산 강제 형변환 적용(Float에서 int64로)

+++ Source/AGSD/Enemy1AIController.cpp
@@ -410,7 +410,7 @@ void AEnemy1AIController::Attacked(float damage)
     FVector2D ScreenLocation;
     APlayerController* PC = UGameplayStatics::GetPlayerController(this, 0);
     if (Enemy.IsValid()) {
-        Enemy->CurrentHP -= damage;
+        Enemy->CurrentHP -= (int64)damage;
         UE_LOG(LogTemp, Display, TEXT("Damamge : %f CurrentHP : %d Actor : %s"), damage, Enemy->CurrentHP, *Enemy->GetName());
         GetWorld()->SpawnActor<AActor>(Enemy->AttackedEffect, GetCharacter()->GetActorLocation(), GetCharacter()->GetActorRotation());
         if (PC && PC->ProjectWorldLocationToScreen(PawnLocation, ScreenLocation)) {
@@ -463,7 +463,7 @@ void AEnemy1AIController::Attacked(float damage, int chanel)
         if (!GetWorldTimerManager().IsTimerActive(Chanel1TimerHandle))
         {
             GetWorldTimerManager().SetTimer(Chanel1TimerHandle, this, &AEnemy1AIController::Chanel1TimerEnd, 0.1f, false);
-            Enemy->CurrentHP -= damage;
+            Enemy->CurrentHP -= (int64)damage;
             if (PC && PC->ProjectWorldLocationToScreen(PawnLocation, ScreenLocation)) {
                 ShowDamage(damage, ScreenLocation);
             }
@@ -473,7 +473,7 @@ void AEnemy1AIController::Attacked(float damage, int chanel)
         if (!GetWorldTimerManager().IsTimerActive(Chanel2TimerHandle))
         {
             GetWorldTimerManager().SetTimer(Chanel2TimerHandle, this, &AEnemy1AIController::Chanel2TimerEnd, 0.1f, false);
-            Enemy->CurrentHP -= damage;
+            Enemy->CurrentHP -= (int64)damage;
             if (PC && PC->ProjectWorldLocationToScreen(PawnLocation, ScreenLocation)) {
                 ShowDamage(damage, ScreenLocation);
             }
@@ -483,7 +483,7 @@ void AEnemy1AIController::Attacked(float damage, int chanel)
         if (!GetWorldTimerManager().IsTimerActive(Chanel3TimerHandle))
         {
             GetWorldTimerManager().SetTimer(Chanel3TimerHandle, this, &AEnemy1AIController::Chanel3TimerEnd, 0.1f, false);
-            Enemy->CurrentHP -= damage;
+            Enemy->CurrentHP -= (int64)damage;
             if (PC && PC->ProjectWorldLocationToScreen(PawnLocation, ScreenLocation)) {
                 ShowDamage(damage, ScreenLocation);
             }
@@ -493,7 +493,7 @@ void AEnemy1AIController::Attacked(float damage, int chanel)
         if (!GetWorldTimerManager().IsTimerActive(Chanel4TimerHandle))
         {
             GetWorldTimerManager().SetTimer(Chanel4TimerHandle, this, &AEnemy1AIController::Chanel4TimerEnd, 0.1f, false);
-            Enemy->CurrentHP -= damage;
+            Enemy->CurrentHP -= (int64)damage;
             if (PC && PC->ProjectWorldLocationToScreen(PawnLocation, ScreenLocation)) {
                 ShowDamage(damage, ScreenLocation);
             }
@@ -503,14 +503,14 @@ void AEnemy1AIController::Attacked(float damage, int chanel)
         if (!GetWorldTimerManager().IsTimerActive(Chanel5TimerHandle))
         {
             GetWorldTimerManager().SetTimer(Chanel5TimerHandle, this, &AEnemy1AIController::Chanel5TimerEnd, 0.1f, false);
-            Enemy->CurrentHP -= damage;
+            Enemy->CurrentHP -= (int64)damage;
             if (PC && PC->ProjectWorldLocationToScreen(PawnLocation, ScreenLocation)) {
                 ShowDamage(damage, ScreenLocation);
             }
         }
         break;
     default:
-        Enemy->CurrentHP -= damage;
+        Enemy->CurrentHP -= (int64)damage;
         if (PC && PC->ProjectWorldLocationToScreen(PawnLocation, ScreenLocation)) {
             ShowDamage(damage, ScreenLocation);
         }
